Player Model and Main Menu
Menu can go to New Game or Jobs
