+++ Assets/Models/Patriot.Type_1.FBX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a6e23941385dd614895e5a38a0ba3946
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
