+++ Assets/Prefabs/Patriot.Type_1.FBX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 14280957bc2a8f9488331ea21203c9f1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
