+++ Assets/Prefabs/Player.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3ff8fec557e7b334ca0564af70c0ab6f
+guid: 8a523362c88bfcd419e0ae0c2110eceb
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
