+++ ProjectSettings/EditorBuildSettings.asset
@@ -4,5 +4,14 @@
 EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
-  m_Scenes: []
+  m_Scenes:
+  - enabled: 1
+    path: Assets/Scenes/MainMenu.unity
+    guid: bf6690e3f616ca54e99fa66b4c61781e
+  - enabled: 1
+    path: Assets/Scenes/Harbor.unity
+    guid: 48251373c6bc0d44f9870ba3d798e774
+  - enabled: 1
+    path: Assets/Scenes/SkyBoxScene.unity
+    guid: 9fc0d4010bbf28b4594072e72b8655ab
   m_configObjects: {}
